refactored site parser modules to use common util module, pep8 compliant for the most part
